Refactor Image Format
